+++ src/EMerchantPay/Libraries/genesis/src/Genesis/Config.php
@@ -42,7 +42,7 @@ final class Config
     /**
      * Library Version
      */
-    const VERSION = '1.18.8';
+    const VERSION = '1.19.2';
 
     /**
      * Core configuration settings